[JS] [Service Bus] package.json in the samples folder
The samples in the `samples` folder fail to run because they require users to install `@azure/service-bus` while they are inside of the source code of `@azure/service-bus`. NPM does this distinction by finding the nearest `package.json`. A rather simple fix for this is to add a `package.json` inside of the `samples` folder.

Closes #3063