Removed the dependencies
The examples have more than one dependency. Instead of providing them all in the package.json, I'm thinking it's probably better to ask users to install them inside of the comments of each file. If it's decided that we want all the dependencies to be listed here, I will do so, of course 🌞